requirements.txt
streamlit
pandas
plotly
openpyxl
scikit-learn
scipy
numpy
streamlit-option-menu
google-genai
python-dotenv
tabulate
